pull jenkins user structs into sep file
